requirements.txt
openai==0.28
ollama
asyncio
websockets
python-dotenv
